requirements.txt
Flask
Flask-Session
psycopg2
SQLAlchemy
werkzeug.security
functools
